shine-HA: add server failover support with fencing
This patch adds basic Lustre HA feature to shine-HA. It is now able to
perform fencing of a server having all targets in errors and lnet
unreachable (eg.  server crash). If the fence command is successful,
HACore will perform a failover of the targets through the shine API, of
both resident and non-resident targets on the selected failover servers.